requirements.txt
pandas
flask
flask-sqlalchemy
flask-login
flask-migrate
flask-wtf
werkzeug
python-dotenv
psycopg2-binary
